requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
python-dotenv==1.0.0
dnspython==2.4.2
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
